@@ -44,30 +44,28 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<java.version>1.8</java.version>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<java.version>11</java.version>
 
-		<lac3-core.version>0.4.9</lac3-core.version>
-		<lac3-cache.version>0.4.9</lac3-cache.version>
-		<lac3-handshake.version>0.4.9</lac3-handshake.version>
-		<lac3-web.version>0.4.9</lac3-web.version>
+		<lac3-core.version>0.6.3</lac3-core.version>
+		<lac3-cache.version>0.6.3</lac3-cache.version>
+		<lac3-handshake.version>0.6.3</lac3-handshake.version>
+		<lac3-web.version>0.6.3</lac3-web.version>
 
-		<lac3-um.version>0.2.7</lac3-um.version>
+		<lac3-um.version>0.3.2</lac3-um.version>
 
-		<dubbo.version>2.7.3</dubbo.version>
-		<curator.version>4.2.0</curator.version>
-		<dubbo-spring-boot.version>2.7.3</dubbo-spring-boot.version>
-		<zookeeper.version>3.4.14</zookeeper.version>
+		<dubbo.version>2.7.7</dubbo.version>
+		<dubbo-spring-boot.version>2.7.7</dubbo-spring-boot.version>
 		<springloaded.version>1.2.8.RELEASE</springloaded.version>
-		<mybatis-spring-boot.version>2.1.0</mybatis-spring-boot.version>
-		<pagehelper-spring-boot.version>1.2.12</pagehelper-spring-boot.version>
-		<druid.version>1.1.20</druid.version>
-		<fastjson.version>1.2.70</fastjson.version>
+		<mybatis-spring-boot.version>2.1.3</mybatis-spring-boot.version>
+		<pagehelper-spring-boot.version>1.3.0</pagehelper-spring-boot.version>
+		<druid.version>1.1.23</druid.version>
+		<fastjson.version>1.2.73</fastjson.version>
 		<attoparser.version>2.0.5.RELEASE</attoparser.version>
 		<unbescape.version>1.1.6.RELEASE</unbescape.version>
-		<rocketmq.version>4.5.2</rocketmq.version>
-		<ognl.version>3.2.11</ognl.version>
+		<rocketmq.version>4.7.1</rocketmq.version>
+		<ognl.version>3.2.14</ognl.version>
 		<easypoi.version>2.4.0</easypoi.version>
 		<jxl.version>2.4.2</jxl.version>
 		<protobuf-java.version>3.10.0</protobuf-java.version>
@@ -100,14 +98,14 @@
 			<artifactId>fastjson</artifactId>
 			<version>${fastjson.version}</version>
 		</dependency>
-		<dependency>
+		<!-- dependency>
 			<groupId>com.sun</groupId>
 			<artifactId>tools</artifactId>
 			<version>${java.version}</version>
 			<scope>system</scope>
 			<systemPath>${env.JAVA_HOME}/lib/tools.jar</systemPath>
 			<optional>true</optional>
-		</dependency>
+		</dependency -->
 	</dependencies>
 
 </project>
