@@ -1,8 +1,8 @@
 package com.linkallcloud.sso.activity;
 
-import com.linkallcloud.core.activity.IActivity;
+import com.linkallcloud.core.activity.ILacLogActivity;
 import com.linkallcloud.sso.domain.LacWebLog;
 
-public interface ILacWebLogActivity extends IActivity<LacWebLog> {
+public interface ILacWebLogActivity extends ILacLogActivity<LacWebLog> {
 
 }
