@@ -1,8 +1,8 @@
 package com.linkallcloud.sso.domain;
 
-import com.linkallcloud.core.laclog.WebBusiLog;
+import com.linkallcloud.core.laclog.BusiLog;
 
-public class LacWebLog extends WebBusiLog {
+public class LacWebLog extends BusiLog {
 	private static final long serialVersionUID = 328977283400771563L;
 
 }
