@@ -1,8 +1,8 @@
 package com.linkallcloud.sso.manager;
 
-import com.linkallcloud.core.manager.IWebBusiLogManager;
+import com.linkallcloud.core.manager.ILacLogManager;
 import com.linkallcloud.sso.domain.LacWebLog;
 
-public interface ILacWebLogManager extends IWebBusiLogManager<LacWebLog> {
+public interface ILacWebLogManager extends ILacLogManager<LacWebLog> {
 
 }
