@@ -1,8 +1,8 @@
 package com.linkallcloud.sso.service;
 
-import com.linkallcloud.core.service.IWebBusiLogService;
+import com.linkallcloud.core.service.ILacLogService;
 import com.linkallcloud.sso.domain.LacWebLog;
 
-public interface ILacWebLogService extends IWebBusiLogService<LacWebLog> {
+public interface ILacWebLogService extends ILacLogService<LacWebLog> {
 
 }
