@@ -1,6 +1,7 @@
 package com.linkallcloud.sso.server;
 
 import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
@@ -12,13 +13,16 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 import com.alibaba.fastjson.support.config.FastJsonConfig;
 import com.alibaba.fastjson.support.spring.FastJsonHttpMessageConverter;
+import com.linkallcloud.core.log.Log;
+import com.linkallcloud.core.log.Logs;
 
 @SpringBootApplication(scanBasePackages = { "com.linkallcloud.sso.redis", "com.linkallcloud.sso.ticket",
 		"com.linkallcloud.sso.server" })
 @EnableTransactionManagement
 @EnableScheduling
 @MapperScan(basePackages = "com.linkallcloud.sso.server.dao")
-public class ServerApplication {
+public class ServerApplication implements CommandLineRunner {
+	private static final Log log = Logs.get();
 
 	@Bean
 	public HttpMessageConverters fastJsonHttpMessageConverters() {
@@ -37,6 +41,11 @@ public class ServerApplication {
 
 	public static void main(String[] args) {
 		new SpringApplicationBuilder(ServerApplication.class).web(WebApplicationType.NONE).run(args);
+	}
+
+	@Override
+	public void run(String... args) throws Exception {
+		log.error("Track:ServerApplication服务器启动完成!");
 	}
 
 }
