@@ -1,6 +1,6 @@
 package com.linkallcloud.sso.server.manager;
 
-import org.apache.dubbo.config.annotation.Service;
+import org.apache.dubbo.config.annotation.DubboService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -10,7 +10,7 @@ import com.linkallcloud.sso.domain.BlackHis;
 import com.linkallcloud.sso.manager.IBlackHisManager;
 import com.linkallcloud.sso.service.IBlackHisService;
 
-@Service(interfaceClass = IBlackHisManager.class, version = "${dubbo.service.version}")
+@DubboService(interfaceClass = IBlackHisManager.class, version = "${dubbo.service.version}")
 @Component
 @Module(name = "黑名单历史")
 public class BlackHisManager extends BaseManager<BlackHis, IBlackHisService> implements IBlackHisManager {
