@@ -1,6 +1,6 @@
 package com.linkallcloud.sso.server.manager;
 
-import org.apache.dubbo.config.annotation.Service;
+import org.apache.dubbo.config.annotation.DubboService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.linkallcloud.core.busilog.annotation.Module;
@@ -8,7 +8,7 @@ import com.linkallcloud.sso.domain.KhAccount;
 import com.linkallcloud.sso.manager.IKhAccountManager;
 import com.linkallcloud.sso.service.IKhAccountService;
 
-@Service(interfaceClass = IKhAccountManager.class, version = "${dubbo.service.version}")
+@DubboService(interfaceClass = IKhAccountManager.class, version = "${dubbo.service.version}")
 @Module(name = "客户账号")
 public class KhAccountManager extends AccountManager<KhAccount, IKhAccountService> implements IKhAccountManager {
 
