@@ -1,6 +1,6 @@
 package com.linkallcloud.sso.server.manager;
 
-import org.apache.dubbo.config.annotation.Service;
+import org.apache.dubbo.config.annotation.DubboService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -10,7 +10,7 @@ import com.linkallcloud.sso.domain.LockHis;
 import com.linkallcloud.sso.manager.ILockHisManager;
 import com.linkallcloud.sso.service.ILockHisService;
 
-@Service(interfaceClass = ILockHisManager.class, version = "${dubbo.service.version}")
+@DubboService(interfaceClass = ILockHisManager.class, version = "${dubbo.service.version}")
 @Component
 @Module(name = "锁/解锁历史")
 public class LockHisManager extends BaseManager<LockHis, ILockHisService> implements ILockHisManager {
