@@ -1,6 +1,6 @@
 package com.linkallcloud.sso.server.manager;
 
-import org.apache.dubbo.config.annotation.Service;
+import org.apache.dubbo.config.annotation.DubboService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -12,7 +12,7 @@ import com.linkallcloud.sso.domain.LoginHis;
 import com.linkallcloud.sso.manager.ILoginHisManager;
 import com.linkallcloud.sso.service.ILoginHisService;
 
-@Service(interfaceClass = ILoginHisManager.class, version = "${dubbo.service.version}")
+@DubboService(interfaceClass = ILoginHisManager.class, version = "${dubbo.service.version}")
 @Component
 @Module(name = "登录日志")
 public class LoginHisManager extends BaseManager<LoginHis, ILoginHisService> implements ILoginHisManager {
