@@ -14,10 +14,15 @@ import com.linkallcloud.core.dto.Trace;
 import com.linkallcloud.core.face.message.request.ListFaceRequest;
 import com.linkallcloud.core.security.MsgSignObject;
 import com.linkallcloud.sso.oapi.kiss.um.ApplicationKiss;
+import com.linkallcloud.sso.oapi.kiss.um.KhUserKiss;
+import com.linkallcloud.sso.oapi.kiss.um.YwUserKiss;
 import com.linkallcloud.um.domain.sys.Application;
 import com.linkallcloud.web.face.FaceAspect;
 import com.linkallcloud.web.face.processor.PackageJsonProcessor;
 import com.linkallcloud.web.face.processor.PackageXmlProcessor;
+import com.linkallcloud.web.face.processor.SimpleJsonProcessor;
+import com.linkallcloud.web.session.SessionUser;
+import com.linkallcloud.web.session.SimpleSessionUser;
 
 @Aspect
 @Component
@@ -26,6 +31,12 @@ public class LacOapiServerFaceAspect extends FaceAspect {
 
 	@Autowired
 	private ApplicationKiss applicationKiss;
+
+	@Autowired
+	private YwUserKiss ywUserKiss;
+
+	@Autowired
+	private KhUserKiss khUserKiss;
 
 	@Pointcut("execution(public * com.linkallcloud.sso.oapi.face..*.*(..))")
 	public void oapiface() {
@@ -39,7 +50,18 @@ public class LacOapiServerFaceAspect extends FaceAspect {
 
 	@Override
 	protected PackageJsonProcessor getPackageJsonProcessor() {
-		PackageJsonProcessor pjp = new PackageJsonProcessor();
+		PackageJsonProcessor pjp = new PackageJsonProcessor() {
+
+			@Override
+			protected SessionUser getSessionUserBySimpleSessionUser(SimpleSessionUser ssu) {
+				if (ssu.getUserType().equals("Yw")) {
+					return ywUserKiss.assembleSessionUser(new Trace(), ssu.getLoginName(), ssu.appCode(), ssu.getOrg());
+				} else {
+					return khUserKiss.assembleSessionUser(new Trace(), ssu.getLoginName(), ssu.appCode(), ssu.getOrg());
+				}
+			}
+
+		};
 		List<Application> apps = findApplications();
 		if (apps != null && !apps.isEmpty()) {
 			for (Application app : apps) {
@@ -59,6 +81,22 @@ public class LacOapiServerFaceAspect extends FaceAspect {
 
 	private List<Application> findApplications() {
 		return applicationKiss.find(new Trace(true), new ListFaceRequest());
+	}
+
+	@Override
+	protected SimpleJsonProcessor getSimpleJsonProcessor() {
+		return new SimpleJsonProcessor() {
+
+			@Override
+			protected SessionUser getSessionUserBySimpleSessionUser(SimpleSessionUser ssu) {
+				if (ssu.getUserType().equals("Yw")) {
+					return ywUserKiss.assembleSessionUser(new Trace(), ssu.getLoginName(), ssu.appCode(), ssu.getOrg());
+				} else {
+					return khUserKiss.assembleSessionUser(new Trace(), ssu.getLoginName(), ssu.appCode(), ssu.getOrg());
+				}
+			}
+
+		};
 	}
 
 }
