@@ -6,6 +6,7 @@ import javax.servlet.MultipartConfigElement;
 
 import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
@@ -25,6 +26,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import com.alibaba.fastjson.serializer.SerializerFeature;
 import com.alibaba.fastjson.support.config.FastJsonConfig;
 import com.alibaba.fastjson.support.spring.FastJsonHttpMessageConverter;
+import com.linkallcloud.core.log.Log;
+import com.linkallcloud.core.log.Logs;
 import com.linkallcloud.sso.pc.aop.LoginFilter;
 import com.linkallcloud.web.interceptors.LacEnvInterceptor;
 import com.linkallcloud.web.support.AppVisitorArgumentResolver;
@@ -33,7 +36,8 @@ import com.linkallcloud.web.support.TraceArgumentResolver;
 @EnableDubbo(multipleConfig = true)
 @Configuration
 @SpringBootApplication(scanBasePackages = { "com.linkallcloud.sso.pc" })
-public class WebPcApplication implements WebMvcConfigurer {
+public class WebPcApplication implements WebMvcConfigurer, CommandLineRunner {
+	private static final Log log = Logs.get();
 
 	@Value("${lac.static.server}")
 	private String staticServer;
@@ -120,6 +124,11 @@ public class WebPcApplication implements WebMvcConfigurer {
 		// 设置总上传数据总大小
 		factory.setMaxRequestSize(DataSize.parse("102400KB"));
 		return factory.createMultipartConfig();
+	}
+
+	@Override
+	public void run(String... args) throws Exception {
+		log.error("Track:WebPcApplication服务器启动完成!");
 	}
 
 }
