@@ -1,13 +1,18 @@
 package com.linkallcloud.sso.pc.aop;
 
-import org.apache.dubbo.config.annotation.Reference;
+import org.apache.dubbo.config.annotation.DubboReference;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Pointcut;
+import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
+import com.alibaba.fastjson.JSON;
+import com.linkallcloud.core.dto.Trace;
+import com.linkallcloud.log.core.rocketmq.RocketmqProducerClient;
 import com.linkallcloud.sso.domain.LacWebLog;
 import com.linkallcloud.sso.manager.ILacWebLogManager;
 import com.linkallcloud.web.busilog.BusiWebLogAspect;
@@ -15,14 +20,38 @@ import com.linkallcloud.web.busilog.BusiWebLogAspect;
 @Aspect
 @Component
 @Order(3)
-public class LacWebLogAspect extends BusiWebLogAspect<LacWebLog, ILacWebLogManager> {
+public class LacWebLogAspect extends BusiWebLogAspect<LacWebLog> {
 
-	@Reference(version = "${dubbo.service.version}", application = "${dubbo.application.id}")
+	@Value("${log.storage.type:es}")
+	private String logStorageType;
+	@Value("${log.appName}")
+	private String appName;
+	@Value("${log.appType}")
+	private String appType;
+
+	@DubboReference(version = "${dubbo.service.version}", application = "${dubbo.application.id}")
 	private ILacWebLogManager lacWebLogManager;
 
 	@Override
-	protected ILacWebLogManager logService() {
-		return lacWebLogManager;
+	protected void logStorage(LacWebLog operatelog) throws Exception {
+		if (operatelog != null) {
+			operatelog.setAppName(appName);
+			operatelog.setAppType(appType);
+			if ("es".equals(logStorageType)) {
+				LacWebLog log = new LacWebLog();
+				BeanUtils.copyProperties(operatelog, log);
+				log.setError(null);
+				log.setCreateTime(null);
+				log.setUuid(null);
+				String logStr = JSON.toJSONString(log);
+				RocketmqProducerClient.getInstance().sendMsg(logStr);
+			} else {
+				if (operatelog.getErrorMessage() != null && operatelog.getErrorMessage().length() > 512) {
+					operatelog.setErrorMessage(operatelog.getErrorMessage().substring(0, 512));
+				}
+				lacWebLogManager.insert(new Trace(operatelog.getTid()), operatelog);
+			}
+		}
 	}
 
 	// @Pointcut("@annotation(com.linkallcloud.core.busilog.annotation.WebLog)")
