@@ -12,16 +12,16 @@ import com.linkallcloud.core.face.message.request.ListFaceRequest;
 import com.linkallcloud.core.face.message.request.ObjectFaceRequest;
 import com.linkallcloud.core.face.message.response.ObjectFaceResponse;
 import com.linkallcloud.core.www.utils.HttpClientFactory;
-import com.linkallcloud.sso.portal.kiss.BaseKiss;
+import com.linkallcloud.sso.portal.kiss.UmBaseKiss;
 import com.linkallcloud.um.domain.sys.Area;
 import com.linkallcloud.um.face.area.ParentCodeAreaRequest;
 
 @Component
-public class AreaKiss extends BaseKiss {
+public class AreaKiss extends UmBaseKiss {
 
 	public List<Tree> getTreeNodes(Trace t, Boolean valid) {
 		String sendMsgPkg = packMessage(t, new ObjectFaceRequest<Boolean>(valid));
-		String responseJson = HttpClientFactory.me(false).post(umBaseUrl + "/face/Area/getTreeNodes", sendMsgPkg);
+		String responseJson = HttpClientFactory.me(false).post(getOapiUrl() + "/face/Area/getTreeNodes", sendMsgPkg);
 		List<Tree> result = unpackMessage(responseJson, new TypeReference<ObjectFaceResponse<List<Tree>>>() {
 		});
 		return result;
@@ -29,7 +29,7 @@ public class AreaKiss extends BaseKiss {
 
 	public Area fetchById(Trace t, Long id) {
 		String sendMsgPkg = packMessage(t, new IdFaceRequest(id, null));
-		String responseJson = HttpClientFactory.me(false).post(umBaseUrl + "/face/Area/fetchById", sendMsgPkg);
+		String responseJson = HttpClientFactory.me(false).post(getOapiUrl() + "/face/Area/fetchById", sendMsgPkg);
 		Area result = unpackMessage(responseJson, new TypeReference<ObjectFaceResponse<Area>>() {
 		});
 		return result;
@@ -37,7 +37,7 @@ public class AreaKiss extends BaseKiss {
 
 	public Area fetchByGovCode(Trace t, String govCode) {
 		String sendMsgPkg = packMessage(t, govCode);
-		String responseJson = HttpClientFactory.me(false).post(umBaseUrl + "/face/Area/fetchByGovCode", sendMsgPkg);
+		String responseJson = HttpClientFactory.me(false).post(getOapiUrl() + "/face/Area/fetchByGovCode", sendMsgPkg);
 		Area result = unpackMessage(responseJson, new TypeReference<ObjectFaceResponse<Area>>() {
 		});
 		return result;
@@ -45,7 +45,7 @@ public class AreaKiss extends BaseKiss {
 
 	public List<Area> find(Trace t, ListFaceRequest req) {
 		String sendMsgPkg = packMessage(t, req);
-		String responseJson = HttpClientFactory.me(false).post(umBaseUrl + "/face/Area/find", sendMsgPkg);
+		String responseJson = HttpClientFactory.me(false).post(getOapiUrl() + "/face/Area/find", sendMsgPkg);
 		List<Area> result = unpackMessage(responseJson, new TypeReference<ObjectFaceResponse<List<Area>>>() {
 		});
 		return result;
@@ -53,7 +53,7 @@ public class AreaKiss extends BaseKiss {
 
 	public List<Area> findChildren(Trace t, ParentCodeAreaRequest req) {
 		String sendMsgPkg = packMessage(t, req);
-		String responseJson = HttpClientFactory.me(false).post(umBaseUrl + "/face/Area/findChildren", sendMsgPkg);
+		String responseJson = HttpClientFactory.me(false).post(getOapiUrl() + "/face/Area/findChildren", sendMsgPkg);
 		List<Area> result = unpackMessage(responseJson, new TypeReference<ObjectFaceResponse<List<Area>>>() {
 		});
 		return result;
